[HIB-102] Browses are now packaged up along with zip archive
The zip files are named based on the granule and the job, and the
browses are named based on what their names already were. The zips will
possibly need to be named by the user code to allow for caching based of
parameters used to generate them which can require knowledge on the
per-process scope.