Remove cidr range from elastic beanstalk
ElasticBeanstalk automatically creates a security group that lets
all traffic through, so ip can't be filted using sg. A NetworkACL can't
be used because for elastic beanstalk to create it needs access to
the outside world.

It work adding the Ip to the NetworkACL after ElasticBeanstalk was
created, but this is a manual step and should be avoided if at all
possible.